+++ src/components/TaskResults/TaskResults.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/nextjs";
 import TaskResults from "./TaskResults";
-import { mockTasks } from "@/actions/__mocks__/getTasks";
+import { mockTasks } from "@/actions/task/__mocks__/getTasks";
 
 const meta: Meta<typeof TaskResults> = {
   title: "Components/TaskResults",
